+++ litassist/commands/updatefacts/__init__.py
@@ -1,0 +1,5 @@
+"""updatefacts command package."""
+
+from .core import updatefacts
+
+__all__ = ["updatefacts"]